[webdriver-client] Move module classes for WebDriver BiDi to their own directory.
Differential Revision: https://phabricator.services.mozilla.com/D131604

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1693839
gecko-commit: 0dce12e803e358de869fb2c9eb068ee34af256fd
gecko-reviewers: webdriver-reviewers, jgraham, jdescottes
