+++ tools/webdriver/webdriver/bidi/modules/__init__.py
@@ -1,0 +1,1 @@
+from .session import Session
